services/market-data/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
yfinance